+++ R/compare.R
@@ -9,11 +9,14 @@
 #' @param table_a A data frame
 #' @param table_b A data frame
 #' @param by <[`tidy-select`][versus_tidy_select]>. Selection of columns to use when matching rows between
-#' \code{.data_a} and \code{.data_b}. Both data frames must be unique on \code{by}.
+#' \code{table_a} and \code{table_b}. Both data frames must be unique on \code{by}.
 #' @param allow_both_NA Logical. If \code{TRUE} a missing value in both data frames is
 #' considered as equal
 #' @param coerce Logical. If \code{FALSE} and columns from the input tables have
 #' differing classes, the function throws an error.
+#' @param table_id A character vector of length 2 providing custom identifiers for
+#' \code{table_a} and \code{table_b} respectively. These identifiers are used in the
+#' output instead of the default "a" and "b".
 #'
 #' @return
 #' \describe{
@@ -58,22 +61,29 @@
 
 #' @rdname compare
 #' @export
-compare <- function(table_a, table_b, by, allow_both_NA = TRUE, coerce = TRUE) {
+compare <- function(
+    table_a,
+    table_b,
+    by,
+    allow_both_NA = TRUE,
+    coerce = TRUE,
+    table_id = c("a", "b")) {
   check_required(by)
   by <- enquo(by)
+  table_id <- clean_table_id(table_id)
   table_chr <- names(enquos(table_a, table_b, .named = TRUE))
   validate_tables(table_a, table_b, coerce = coerce)
 
   by_names <- get_by_names(table_a, table_b, by = by)
 
   table_summ <- tibble(
-    table = c("table_a", "table_b"),
+    table = table_id,
     expr = table_chr,
     nrow = c(nrow(table_a), nrow(table_b)),
     ncol = c(ncol(table_a), ncol(table_b))
   )
 
-  tbl_contents <- get_contents(table_a, table_b, by = by_names)
+  tbl_contents <- get_contents(table_a, table_b, by = by_names, table_id = table_id)
 
   matches <- withCallingHandlers(
     locate_matches(table_a, table_b, by = by_names),
@@ -87,7 +97,8 @@ compare <- function(table_a, table_b, by, allow_both_NA = TRUE, coerce = TRUE) {
     table_a,
     table_b,
     by = by_names,
-    matches = matches
+    matches = matches,
+    table_id = table_id
   )
 
   tbl_contents$compare$diff_rows <- tbl_contents$compare$column %>%
@@ -95,21 +106,22 @@ compare <- function(table_a, table_b, by, allow_both_NA = TRUE, coerce = TRUE) {
       table_a = table_a,
       table_b = table_b,
       matches = matches,
-      allow_both_NA = allow_both_NA
+      allow_both_NA = allow_both_NA,
+      table_id = table_id
     )
 
   tbl_contents$compare <- tbl_contents$compare %>%
     mutate(n_diffs = map_int(diff_rows, nrow), .after = column)
 
-  out <- list(
+  list(
     tables = table_summ,
     by = tbl_contents$by,
     intersection = tbl_contents$compare,
     unmatched_cols = tbl_contents$unmatched_cols,
     unmatched_rows = unmatched_rows,
-    input = store_tables(table_a, table_b)
-  )
-  structure(out, class = "vs_comparison")
+    input = store_tables(table_a, table_b, table_id)
+  ) %>%
+    structure(class = "vs_comparison")
 }
 
 # Methods -----------
@@ -175,18 +187,18 @@ split_matches <- function(matches) {
   )
 }
 
-get_unmatched_rows <- function(table_a, table_b, by, matches) {
-  unmatched <- list(
-    a = fsubset(table_a, matches$a, by),
-    b = fsubset(table_b, matches$b, by)
-  )
-  unmatched %>%
+get_unmatched_rows <- function(table_a, table_b, by, matches, table_id) {
+  list(
+    fsubset(table_a, matches$a, by),
+    fsubset(table_b, matches$b, by)
+  ) %>%
+    setNames(table_id) %>%
     bind_rows(.id = "table") %>%
     mutate(row = with(matches, c(a, b))) %>%
     as_tibble()
 }
 
-converge <- function(table_a, table_b, by, matches) {
+converge <- function(table_a, table_b, by, matches, table_id) {
   common_cols <- setdiff(intersect(names(table_a), names(table_b)), by)
 
   by_a <- fsubset(table_a, matches$common$a, by)
@@ -195,20 +207,25 @@ converge <- function(table_a, table_b, by, matches) {
 
   add_vars(
     by_a,
-    frename(common_a, \(nm) paste0(nm, "_a")),
-    frename(common_b, \(nm) paste0(nm, "_b"))
+    frename(common_a, \(nm) paste0(nm, "_", table_id[1])),
+    frename(common_b, \(nm) paste0(nm, "_", table_id[2]))
   )
 }
 
-join_split <- function(table_a, table_b, by) {
+join_split <- function(table_a, table_b, by, table_id) {
   matches <- locate_matches(table_a, table_b, by)
-  intersection <- converge(table_a, table_b, by, matches)
-  unmatched_rows <- get_unmatched_rows(table_a, table_b, by, matches)
+  intersection <- converge(table_a, table_b, by, matches, table_id)
+  unmatched_rows <- get_unmatched_rows(table_a, table_b, by, matches, table_id)
   list(intersection = intersection, unmatched_rows = unmatched_rows)
 }
 
-get_contents <- function(table_a, table_b, by) {
-  tbl_contents <- join_split(contents(table_a), contents(table_b), by = "column")
+get_contents <- function(table_a, table_b, by, table_id) {
+  tbl_contents <- join_split(
+    contents(table_a),
+    contents(table_b),
+    by = "column",
+    table_id = table_id
+  )
   out <- list()
 
   out$by <- tbl_contents$intersection %>%
@@ -223,9 +240,9 @@ get_contents <- function(table_a, table_b, by) {
   out
 }
 
-store_tables <- function(table_a, table_b) {
+store_tables <- function(table_a, table_b, table_id) {
   env <- new_environment()
-  env$value <- list(a = table_a, b = table_b)
+  env$value <- list(table_a, table_b) %>% setNames(table_id)
   dt_copy <- getOption("versus.copy_data_table", default = FALSE)
   if (dt_copy) {
     env$value <- env$value %>%
@@ -255,6 +272,30 @@ rethrow_match_relationship <- function(table_a, table_b, by) {
 
     cli_abort(c(top_msg, info, itemize_row(tbl_row)), call = call)
   }
+}
+
+clean_table_id <- function(table_id, call = caller_env()) {
+  if (!is_character(table_id, n = 2)) {
+    message <- c(
+      "{.arg table_id} must be a character vector of length 2",
+      i = "{.arg table_id} is {.obj_type_friendly {table_id}} of length {length(table_id)}"
+    )
+    cli_abort(message, call = call)
+  }
+  attributes(table_id) <- NULL
+  new <- table_id %>%
+    vec_as_names(repair = "universal", quiet = TRUE) %>%
+    # second vec_as_names() is needed due to vctrs issue #1013
+    vec_as_names(repair = "unique", quiet = TRUE)
+  old <- table_id %|% ""
+  is_changed <- new != old
+  if (!any(is_changed)) {
+    return(table_id)
+  }
+  bullets <- paste0("`", old[is_changed], "` -> `", new[is_changed], "`")
+  message <- c("{.arg table_id} has been adjusted", set_names(bullets, "*"))
+  cli_inform(message = message)
+  new
 }
 
 validate_tables <- function(table_a, table_b, coerce, call = caller_env()) {