+++ R/slice-unmatched.R
@@ -17,13 +17,16 @@
 #' # slice_unmatched(comp, "a") output is the same as
 #' example_df_a |> dplyr::anti_join(example_df_b, by = comp$by$column)
 #'
+#' comp <- compare(example_df_a, example_df_b, by = car, table_id = c("old", "new"))
+#' comp |> slice_unmatched("old")
+#'
 #' comp |> slice_unmatched_both()
 
 #' @rdname slice_unmatched
 #' @export
 slice_unmatched <- function(comparison, table) {
   assert_is_comparison(enquo(comparison))
-  assert_table_is_a_or_b(enquo(table))
+  check_table_arg(enquo(table), comparison)
   slice_unmatched_impl(comparison, table)
 }
 
@@ -31,10 +34,12 @@ slice_unmatched <- function(comparison, table) {
 #' @export
 slice_unmatched_both <- function(comparison) {
   assert_is_comparison(enquo(comparison))
+  table_id <- comparison$tables$table
 
   out_cols <- with(comparison, c(by$column, intersection$column))
 
-  c(a = "a", b = "b") %>%
+  table_id %>%
+    setNames(table_id) %>%
     map(slice_unmatched_impl, comparison = comparison, j = out_cols) %>%
     ensure_ptype_compatible() %>%
     bind_rows(.id = "table")